Fixed bugs and changed UUID gen

+++ supabase/functions/delete-content/index.ts
@@ -1,10 +1,22 @@
 // @ts-ignore
 import { serve } from 'std/server';
-import { TableName, connect, deleteData, deleteResponseWrapper } from '../_shared/helpers.ts';
+import {
+  TableName,
+  connect,
+  convertContentTypeToTableName,
+  deleteData,
+  deleteResponseWrapper,
+  fetchData,
+} from '../_shared/helpers.ts';
+import type { ContentType } from '../_shared/content';
 
 serve(async (req: Request) => {
   return await connect(req, async (client, body) => {
-    let { id, tableName } = body as { id: number; tableName: TableName };
+    let { id, type } = body as { id: number; type: TableName | ContentType };
+
+    const convertedType = convertContentTypeToTableName(type as ContentType);
+    const tableName = convertedType ? convertedType : type;
+
     if (
       ![
         'ability_block',
@@ -26,8 +38,24 @@ serve(async (req: Request) => {
       };
     }
 
+    const existingContents = await fetchData(client, tableName as TableName, [
+      { column: 'id', value: id },
+    ]);
+    const existingContent = existingContents.length > 0 ? existingContents[0] : undefined;
+    if (!existingContent) {
+      return {
+        status: 'error',
+        message: 'Content not found',
+      };
+    }
+
+    // Delete any attached trait
+    if (existingContent.trait_id) {
+      await deleteData(client, 'trait', existingContent.trait_id);
+    }
+
     // TODO: Confirm that you can only delete content that you have permission to delete
-    const result = await deleteData(client, tableName, id);
+    const result = await deleteData(client, tableName as TableName, id);
 
     return deleteResponseWrapper(result);
   });
